Move block pos argument stuff into a different folder
